// src/panels/PanelCanvas/types/RectSimple.tsx
export type RectSimple = {
    x: number
    y: number
    width: number
    height: number
    scaleX: number
    scaleY: number
}

export type UnifiedCanvasViewInfos = {
    canvasX: number
    canvasY: number
    viewportPointerX: number
    viewportPointerY: number
    viewPointerX: number
    viewPointerY: number
    scale: number
    isDown: boolean
}